requirements.txt
# Работа с данными
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Kafka
kafka-python==2.0.2

# ML / DL
scikit-learn==1.3.2
joblib==1.3.2

# Визуализация
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Dashboard
streamlit==1.30.0

# Утилиты
python-dotenv==1.0.0
tqdm==4.66.1